+++ src/USER-LB/fix_lb_rigid_pc_sphere.cpp
@@ -1430,7 +1430,7 @@ int FixLbRigidPCSphere::dof(int igroup)
 double FixLbRigidPCSphere::memory_usage()
 {
   int nmax = atom->nmax;
-  double bytes = nmax * sizeof(int);
+  double bytes = (double)nmax * sizeof(int);
   bytes += (double)nmax*3 * sizeof(double);
   bytes += (double)maxvatom*6 * sizeof(double);
