+++ src/USER-MISC/fix_propel_self.cpp
@@ -130,7 +130,7 @@ int FixPropelSelf::setmask()
 double FixPropelSelf::memory_usage()
 {
   // magnitude + thermostat_orient + mode + n_types_filter + apply_to_type
-  double bytes = sizeof(double) + 3*sizeof(int) + sizeof(int*);
+  double bytes = (double)sizeof(double) + 3*sizeof(int) + sizeof(int*);
   bytes += (double)sizeof(int)*atom->ntypes*n_types_filter;
 
   return bytes;
